modified:   ainsuForth.ino 	modified:   ainsuForthsketch.cpp
Fairly good result: the compiler prompt varies by the amount
of inbound traffic waiting, to avoid serious overruns.

It'll spit out an 'ul' ("upload detected") to show it
feels the burden of a code paste-in, up to a point;
and then won't even try to do that (but will still
output formatting line endings, if it can).

When it's truly close to overrun, it will not print
anything except to echo input -- in which case, more
than one line of input appears on a single line of
output.

IF CODE IS typed by hand, that's slow enough to echo
the verbose prompt, which is set to 'compiled' to
match the corresponding behavior in gforth.