modified:   README.md 	modified:   ainsuForth.ino 	modified:   ainsuForthsketch.cpp
Stable code running on Gemma M0 - mostly, this is the snapshot taken today.

Will modify further, from the copy of the snapshot (which is now also the
main program version's code).

The goal is to move forward with making ainsuForth more agnostic to
which target it's running on, without sacrificing features of the
implementation specific to a given target board.